requirements.txt
torch>=1.7.0
torchvision>=0.8.0
numpy>=1.19.0
opencv-python>=4.0.0
Pillow>=8.0.0
matplotlib>=3.3.0
scikit-image>=0.17.0
tqdm>=4.50.0
pyyaml>=5.4.0
timm>=0.6.0
kornia>=0.6.0
ruamel.yaml>=0.17.21
ptflops>=0.6.8
tensorboard>=2.6.0
